move fast_topk
